test(tests): Validation API tests were reorganized
